@@ -20,7 +20,7 @@ publish_to: 'none' # Remove this line if you wish to publish to pub.dev
 version: 1.0.0+1
 
 environment:
-  sdk: ">=2.7.0 <3.0.0"
+  sdk: ">=2.17.5 <3.0.0"
 
 # Dependencies specify other packages that your package needs in order to work.
 # To automatically upgrade your package dependencies to the latest versions
@@ -39,12 +39,12 @@ dependencies:
   flutter_svg: ^1.1.5
   flutter_rating_bar: ^4.0.1
   sidebarx: ^0.14.0
-  http: ^0.12.2
-  carousel_slider: ^3.0.0
-  shimmer: ^2.0.0
-  get_storage: ^2.0.2
-  cached_network_image: ^2.0.0
-  get: ^4.6.5
+  http: ^0.13.5
+  card_swiper: ^2.0.4
+  page_transition: ^2.0.9
+  provider: ^6.0.5
+  flutter_iconly: ^1.0.2
+  fancy_shimmer_image: ^2.0.2
 
 dev_dependencies:
   flutter_test:
