@@ -2,10 +2,11 @@
 import 'package:cocotea_eco/Screen/Bill.dart/Appbarbill.dart';
 import 'package:cocotea_eco/Screen/Bill.dart/Body.dart';
 import 'package:cocotea_eco/Screen/BottomNavBar/BottomNavBar.dart';
+import 'package:cocotea_eco/Screen/Dashboard/Appbar.dart';
 import 'package:flutter/material.dart';
 
 class BillPage extends StatefulWidget {
-  const BillPage({Key key}) : super(key: key);
+  const BillPage({Key? key}) : super(key: key);
 
   @override
   State<BillPage> createState() => _BillPageState();
@@ -17,7 +18,7 @@ class _BillPageState extends State<BillPage> {
     return Scaffold(
       backgroundColor: Colors.white,
       appBar: BillAppbar(context), //menu
-      body: Body(),
+      body: BodyBill(),
       bottomNavigationBar: bottomNavigationBar(),
     );
   }
