@@ -3,14 +3,14 @@ import 'package:cocotea_eco/Screen/Bill.dart/ListBill.dart';
 import 'package:cocotea_eco/Screen/Bill.dart/Statusbill.dart';
 import 'package:flutter/material.dart';
 
-class Body extends StatefulWidget {
-  const Body({Key key}) : super(key: key);
+class BodyBill extends StatefulWidget {
+  const BodyBill({Key? key}) : super(key: key);
 
   @override
-  State<Body> createState() => _BodyState();
+  State<BodyBill> createState() => _BodyState();
 }
 
-class _BodyState extends State<Body> {
+class _BodyState extends State<BodyBill> {
   static const time = '14/10/2022';
   @override
   Widget build(BuildContext context) {
