@@ -3,7 +3,7 @@ import 'package:cocotea_eco/Screen/Dashboard/Dashboard_page.dart';
 import 'package:flutter/material.dart';
 
 class CartPage extends StatefulWidget {
-  const CartPage({Key key}) : super(key: key);
+  const CartPage({Key? key}) : super(key: key);
 
   @override
   State<CartPage> createState() => _CartPageState();
