@@ -7,7 +7,7 @@ import 'package:flutter/material.dart';
 
 class DashboardPage extends StatefulWidget {
   
-  const DashboardPage({Key key}) : super(key: key);
+  const DashboardPage({Key? key}) : super(key: key);
 
   @override
   State<DashboardPage> createState() => _DashboardPageState();
@@ -20,7 +20,7 @@ class _DashboardPageState extends State<DashboardPage> {
       backgroundColor: Colors.white,
       appBar: homeAppBar(context,
       ), //menu
-      body: Body(),
+      body: BodyDB(),
       bottomNavigationBar: bottomNavigationBar(),
     );
   }
