@@ -1,30 +1,91 @@
-import 'package:cocotea_eco/Components/banner_data_found.dart';
-import 'package:cocotea_eco/Controller/banner_controller.dart';
-import 'package:cocotea_eco/Screen/Dashboard/DiscountCard_Loading.dart';
+import 'package:cocotea_eco/Screen/Product/Constant.dart';
 import 'package:flutter/material.dart';
-import 'package:get/get.dart';
 
 class DiscountCard extends StatelessWidget {
-  const DiscountCard({Key key}) : super(key: key);
+  const DiscountCard({Key? key}) : super(key: key);
 
   @override
   Widget build(BuildContext context) {
-    return Scaffold(
-      body: SafeArea(
-          child: GetBuilder<BannerController>(
-            builder: (_c){
-              if(_c.isLoading)
-                if(_c.bannerData.length>0)
-                  return BannerDataFound(_c.bannerData);
-                else
-                  return DiscountCard_Loading();
-              else
-                if(_c.bannerData.length>0)
-                  return BannerDataFound(_c.bannerData);
-                else
-                  return Container();
-            },
-          )
+  
+    Size size = MediaQuery.of(context).size;
+    return Container(
+      width: double.infinity,
+      height: size.height * 0.2,
+       
+      decoration: BoxDecoration(
+        borderRadius: BorderRadius.circular(18),
+        gradient: const LinearGradient(
+            colors: [
+               Color.fromARGB(255, 100, 167, 97),
+              Color.fromARGB(255, 199, 196, 110),
+             
+            ],
+            begin: FractionalOffset(0.0, 0.0),
+            end: FractionalOffset(1.0, 0.0),
+            stops: [0.0, 1.0],
+            tileMode: TileMode.clamp),
+      ),
+      child: Row(
+        children: [
+          Flexible(
+            flex: 2,
+            child: Padding(
+              padding: const EdgeInsets.all(14.0),
+              child: Container(
+                height: double.infinity,
+                decoration: BoxDecoration(
+                  color: Color.fromARGB(255, 98, 143, 105),
+                  borderRadius: BorderRadius.circular(18),
+                ),
+                child: Padding(
+                  padding: const EdgeInsets.all(8.0),
+                  child: Column(
+                    // mainAxisSize: MainAxisSize.max,
+                    // mainAxisAlignment: MainAxisAlignment.center,
+                    // crossAxisAlignment: CrossAxisAlignment.center,
+                    children: const [
+                      Text(
+                        "Ưu đãi hấp dẫn cho thành viên mới",
+                        style: TextStyle(color: Colors.white, fontWeight: FontWeight.bold,fontSize: 14),
+                      ),
+                      SizedBox(
+                        height: 18,
+                      ),
+                      Flexible(
+                        child: SizedBox(
+                          width: double.infinity,
+                          child: FittedBox(
+                            fit: BoxFit.fill,
+                            child: Text(
+                              "50%\nOFF",
+                              style: TextStyle(
+                                color: Colors.white,
+                                fontWeight: FontWeight.bold,
+                                // fontSize: 300,
+                              ),
+                            ),
+                          ),
+                        ),
+                      ),
+                    ],
+                  ),
+                ),
+              ),
+            ),
+          ),
+          Flexible(
+            flex: 3,
+            child: Padding(
+              padding: const EdgeInsets.all(14.0),
+              child: Image.network(
+                "https://f9-zpcloud.zdn.vn/5034820970077684476/02dfc9b39db340ed19a2.jpg",
+                width: double.infinity,
+                // height: double.infinity,
+                fit: BoxFit.contain,
+              ),
+            ),
+          ),
+        ],
       ),
     );
   }
