@@ -4,7 +4,7 @@ import 'package:cocotea_eco/Screen/Detail/Detailpage.dart';
 import 'package:flutter/material.dart';
 class IntemList extends StatelessWidget {
   const IntemList({
-    Key key,
+    Key? key,
   }) : super(key: key);
 
   @override
