@@ -2,7 +2,7 @@ import 'package:flutter/material.dart';
 import 'package:flutter_rating_bar/flutter_rating_bar.dart';
 class titleandrating extends StatelessWidget {
   const titleandrating({
-    Key key,
+    Key? key,
   }) : super(key: key);
 
   @override
