@@ -6,7 +6,7 @@ import 'components/sign_up_top_image.dart';
 import 'components/signup_form.dart';
 
 class SignUpScreen extends StatelessWidget {
-  const SignUpScreen({Key key}) : super(key: key);
+  const SignUpScreen({Key? key}) : super(key: key);
 
   @override
   Widget build(BuildContext context) {
@@ -33,7 +33,7 @@ class SignUpScreen extends StatelessWidget {
                 ),
               )
             ],
-          ),
+          ), 
         ),
       ),
     );
@@ -42,7 +42,7 @@ class SignUpScreen extends StatelessWidget {
 
 class MobileSignupScreen extends StatelessWidget {
   const MobileSignupScreen({
-    Key key,
+    Key? key,
   }) : super(key: key);
 
   @override
