@@ -5,7 +5,7 @@ import '../../Signup/signup_screen.dart';
 
 class LoginAndSignupBtn extends StatelessWidget {
   const LoginAndSignupBtn({
-    Key key,
+    Key? key,
   }) : super(key: key);
 
   @override
