@@ -5,7 +5,7 @@ import '../../../../Product/Constant.dart';
 
 class WelcomeImage extends StatelessWidget {
   const WelcomeImage({
-    Key key,
+    Key? key,
   }) : super(key: key);
 
   @override
