@@ -6,7 +6,7 @@ import 'components/login_signup_btn.dart';
 import 'components/welcome_image.dart';
 
 class WelcomeScreen extends StatelessWidget {
-  const WelcomeScreen({Key key}) : super(key: key);
+  const WelcomeScreen({Key ?key}) : super(key: key);
 
   @override
   Widget build(BuildContext context) {
@@ -33,7 +33,7 @@ class WelcomeScreen extends StatelessWidget {
                 ),
               ],
             ),
-            mobile: const MobileWelcomeScreen(),
+            mobile: const MobileWelcomeScreen(), 
           ),
         ),
       ),
@@ -43,7 +43,7 @@ class WelcomeScreen extends StatelessWidget {
 
 class MobileWelcomeScreen extends StatelessWidget {
   const MobileWelcomeScreen({
-    Key key,
+    Key? key,
   }) : super(key: key);
 
   @override
