@@ -2,14 +2,12 @@ import 'package:flutter/material.dart';
 
 class Responsive extends StatelessWidget {
   final Widget mobile;
-  final Widget tablet;
   final Widget desktop;
 
   const Responsive({
-    Key key,
-     this.mobile,
-    this.tablet,
-     this.desktop,
+     Key? key,
+      required this.mobile,
+      required this.desktop,
   }) : super(key: key);
 
   static bool isMobile(BuildContext context) =>
@@ -27,8 +25,6 @@ class Responsive extends StatelessWidget {
     final Size size = MediaQuery.of(context).size;
     if (size.width > 992) {
       return desktop;
-    } else if (size.width >= 576 && tablet != null) {
-      return tablet;
     } else {
       return mobile;
     }
