@@ -7,7 +7,7 @@ import 'package:flutter/material.dart';
 import 'package:flutter_svg/flutter_svg.dart';
 
 class ProductPage extends StatefulWidget {
-  const ProductPage({Key key}) : super(key: key);
+  const ProductPage({Key? key}) : super(key: key);
 
   @override
   State<ProductPage> createState() => _ProductPageState();
@@ -19,7 +19,7 @@ class _ProductPageState extends State<ProductPage> {
     return Scaffold(
       backgroundColor: Colors.white,
       appBar: homeAppBar(context), //menu
-      body: Body(),
+      body: BodyPD(),
       bottomNavigationBar: bottomNavigationBar(),
     );
   }
