@@ -2,7 +2,7 @@ import 'package:cocotea_eco/Screen/Product/Constant.dart';
 import 'package:flutter/material.dart';
 
 class OrDivider extends StatelessWidget {
-  const OrDivider({Key key}) : super(key: key);
+  const OrDivider({Key ?key}) : super(key: key);
 
   @override
   Widget build(BuildContext context) {
