@@ -6,9 +6,9 @@ class SocalIcon extends StatelessWidget {
   final String iconSrc;
   final Function press;
   const SocalIcon({
-    Key key,
-    this.iconSrc,
-    this.press,
+    Key ?key,
+    required this.iconSrc,
+    required this.press,
   }) : super(key: key);
 
   @override
