@@ -4,7 +4,7 @@ class ProductItem {
   int quantity;
   final Product product;
 
-  ProductItem({this.quantity = 1,  this.product});
+  ProductItem({this.quantity = 1,  required this.product});
 
   void increment() {
     quantity++;
