@@ -6,8 +6,8 @@ import 'package:flutter/material.dart';
 
 class CartDetailsViewCard extends StatelessWidget {
   const CartDetailsViewCard({
-    Key key,
-     this.productItem,
+    Key ?key,
+     required this.productItem,
   }) : super(key: key);
 
   final ProductItem productItem;
@@ -25,7 +25,7 @@ class CartDetailsViewCard extends StatelessWidget {
         productItem.product.title,
         style: Theme.of(context)
             .textTheme
-            .subtitle1
+            .subtitle1!
             .copyWith(fontWeight: FontWeight.bold),
       ),
       trailing: FittedBox(
@@ -36,7 +36,7 @@ class CartDetailsViewCard extends StatelessWidget {
               "  x ${productItem.quantity}",
               style: Theme.of(context)
                   .textTheme
-                  .subtitle1
+                  .subtitle1!
                   .copyWith(fontWeight: FontWeight.bold),
             )
           ],
